pyproject.toml
[tool.poetry]
name = "pythonproject"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
beautifulsoup4 = "^4.13.3"
loguru = "^0.7.3"
selenium = "^4.29.0"
pandas = "^2.2.3"
langdetect = "^1.0.9"
httpx = "^0.28.1"
webdriver-manager = "^4.0.2"
pathvalidate = "^3.2.3"
pydantic = "^2.10.6"
mysql-connector-python = "^9.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
